requirements.txt
python-telegram-bot
transformers